% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.all.numeric.R
\name{is.all.numeric}
\alias{is.all.numeric}
\title{Utility extract function}
\usage{
is.all.numeric(x, what = c("test", "vector"), extras = c(".", "NA", "na",
  "null", "NULL"))
}
\description{
Utility extract function
}
\author{
Dan Fu (not really)
}

