% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string.to.vector.of.words.R
\name{string.to.vector.of.words}
\alias{string.to.vector.of.words}
\title{Utility extract function}
\usage{
string.to.vector.of.words(string)
}
\description{
Utility extract function
}
\author{
Dan Fu
}

